Adjusted auth controls
