+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/components/GetCodeStep.tsx
@@ -3,6 +3,7 @@ import { t } from "ttag";
 import { useUpdateSettingsMutation } from "metabase/api";
 import { CodeEditor } from "metabase/common/components/CodeEditor";
 import { MoreServerSnippetExamplesLink } from "metabase/embedding/components/MoreServerSnippetExamplesLink/MoreServerSnippetExamplesLink";
+import { MetabaseAccountSection } from "metabase/embedding/embedding-iframe-sdk-setup/components/Authentication/MetabseAccountSection";
 import { CopyCodeSnippetButton } from "metabase/embedding/embedding-iframe-sdk-setup/components/CodeSnippet/CopyCodeSnippetButton";
 import { useSdkIframeEmbedServerSnippet } from "metabase/embedding/embedding-iframe-sdk-setup/hooks/use-sdk-iframe-embed-server-snippet";
 import { EmbedServerSnippetLanguageSelect } from "metabase/public/components/EmbedServerSnippetLanguageSelect/EmbedServerSnippetLanguageSelect";
@@ -16,6 +17,8 @@ import { useSdkIframeEmbedSnippet } from "../hooks/use-sdk-iframe-embed-snippet"
 export const GetCodeStep = () => {
   const { experience, resource, settings } = useSdkIframeEmbedSetupContext();
   const [updateInstanceSettings] = useUpdateSettingsMutation();
+
+  const isGuestEmbed = !!settings.isGuest;
 
   const serverSnippetData = useSdkIframeEmbedServerSnippet();
   const snippet = useSdkIframeEmbedSnippet();
@@ -44,6 +47,8 @@ export const GetCodeStep = () => {
 
   return (
     <Stack gap="md">
+      {!isGuestEmbed && <MetabaseAccountSection />}
+
       <Card p="md">
         <Text size="lg" fw="bold" mb="md">
           {t`Embed code`}
