+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/constants.ts
@@ -24,23 +24,27 @@ export const getEmbedExperiences = ({
       value: "dashboard",
       title: t`Dashboard`,
       description: t`Embed an entire dashboard with multiple charts and filters`,
+      supportsGuestEmbed: true,
     },
     {
       value: "chart",
       title: t`Chart`,
       description: t`Embed a single chart`,
+      supportsGuestEmbed: true,
     },
     {
       value: "exploration",
       title: t`Exploration`,
       description: t`Embed an interactive data exploration experience`,
       showUpsell: !isSimpleEmbedFeatureAvailable,
+      supportsGuestEmbed: false,
     },
     {
       value: "browser",
       title: t`Browser`,
       description: t`Embed a browser to manage dashboards and charts`,
       showUpsell: !isSimpleEmbedFeatureAvailable,
+      supportsGuestEmbed: false,
     },
     ...(isMetabotAvailable
       ? [
@@ -49,6 +53,7 @@ export const getEmbedExperiences = ({
             title: t`Metabot`,
             description: t`Embed a Metabot chat interface`,
             showUpsell: !isSimpleEmbedFeatureAvailable,
+            supportsGuestEmbed: false,
           },
         ]
       : []),
@@ -57,6 +62,7 @@ export const getEmbedExperiences = ({
     description: string;
     value: SdkIframeEmbedSetupExperience;
     showUpsell?: boolean;
+    supportsGuestEmbed: boolean;
   }[];
 
 type EmbedStepConfig = {
